requirements.txt
streamlit>=1.35
supabase>=2.6
google-api-python-client>=2.130
google-auth>=2.29
google-auth-oauthlib>=1.2
requests>=2.32
